Rename ViewHelpers to Helpers and remove old implementation.
